Update intracell_analysis_v2.py
Modify reference electrode data to be a dictionary of rates and rate file names, instead of file names for one rate; these changes are necessary for other updates which seek to correct for effective rate in the electrodes and utilize multiple rate data of the electrodes.

Incorporate changes for using alternative error metrics, such as dVdQ instead of the V-Q curve.